Crate problem077

Problem 77 (Prime summations)

Solution detail

An iterative dynamic programming approach is best here.

First, we need an upper bound on the answer, so we know how many primes to consider. An extremely crude upper bound is 3n - we can write this as a sum of 2's and 3's in at least n ways. So to guarantee to find the answer, sieving all primes up to 3n will suffice. To actually find the answer:

Start with an array indexed by 0, 1, 2, ..., 3n, with a[0] = 1 and all other a[i] = 0. Note that initially, a contains the numbers of ways of producing each target 0, 1, 2, ..., 3n using primes taking values from the empty set.

We will iteratively add one prime at a time to the pool of possible primes, updating the array as we do so. Indeed, suppose we add a prime with value p. Then we update a as follows:

for j = p to end of array:
    a[j] += a[j-p]

After doing this for a prime p, we can chop the array off after any index with an entry at least n, because we know the answer will not exceed that index.

After doing this process for all primes p below our upper bound, the first index with an entry over n is the answer.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Finds the first number that can be written as a sum of primes in at least the given number of ways.